judicial_document_management/requirements.txt
asgiref==3.11.0
brotli==1.2.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cobble==0.1.4
cryptography==41.0.5
cssselect2==0.9.0
defusedxml==0.8.0rc2
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
django-templated-mail==1.1.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
djoser==2.2.1
fonttools==4.61.1
html2text==2025.4.15
idna==3.4
lxml==6.0.2
mammoth==1.11.0
oauthlib==3.2.2
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==2.21
pydyf==0.12.1
PyJWT==2.8.0
pyphen==0.17.2
python-docx==1.2.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2023.3
urllib3==2.0.7
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
